Update board_front.js
Update includes two main changes. 
First, it allows for multiple elements in growth (OR or multiple of the same). It does not allow for AND, which isn't' currently used in the game but is a reasonable design choice.
Second, it updates how the picture is scaled. The picture now automatically scales to the size of the window (accounting for the changing Special Rule size). The user can change the image size using the spirit-image-size option that was already in the HTML (but not used, to my knowledge).
Requires a change to Growth options.css as well