Update adversary.css
Adds "Base Difficulty" to the CSS
Adds "Difficulty Level" to the CSS (that's the little number next to each level)
Requires changes to HTML - this seems to only be available in the example. Should we create a folder for the html with no content? I also think an empty folder for folk to add their designs to would be helpful. I had trouble getting started initially.

Updates to the following icons:
Escalation
City
Presence