allowing reverse-named dual lands
I don't expect users to remember if its wetland-jungle or jungle-wetland. This solves that problem.
However, it may cause some issues down the line if these icons are used by name in other places. I've tried to find a way to make one class be interpreted as another, but haven't figured that out. Perhaps I need to solve this in the JS instead of the CSS?